code cleanup in sidebar
