+++ apps/sim/app/workspace/[workspaceId]/w/components/sidebar/components/index.ts
@@ -1,9 +1,12 @@
 export { CreateMenu } from './create-menu/create-menu'
+export { FloatingNavigation } from './floating-navigation/floating-navigation'
 export { FolderTree } from './folder-tree/folder-tree'
 export { HelpModal } from './help-modal/help-modal'
+export { KeyboardShortcut } from './keyboard-shortcut/keyboard-shortcut'
 export { KnowledgeBaseTags } from './knowledge-base-tags/knowledge-base-tags'
 export { KnowledgeTags } from './knowledge-tags/knowledge-tags'
 export { LogsFilters } from './logs-filters/logs-filters'
+export { NavigationItem } from './navigation-item/navigation-item'
 export { SettingsModal } from './settings-modal/settings-modal'
 export { SubscriptionModal } from './subscription-modal/subscription-modal'
 export { Toolbar } from './toolbar/toolbar'
